de: Update whole translation to current version
A big update of the complete german translation. Not perfect in detail
but at least a mostly complete rewrite, removing of various 'fuzzy' and
'unclear' status flags.

Also started to translate new added strings since last *.po file updates.

I switched the german translation to de_DE (from de_CH) as this is the
language of the most german speaking people. In detail this means mostly
switching from '*ss*' to '*ß*' like for the german word for 'must'
(muss -> muß).

And of course some small fixups for spelling and grammer.