Increase cache validity period in nginx.conf
This revised nginx.conf holds the cache for one week, truly.  For some reason there are
two tuning knobs required to make one adjustment (cache duration).  Without this fix the
cache was only held for 10 minutes.  The most important change is the "inactive=1w" which
as I implied, is the *second* reference to one week.  Sorry for this oversight.

Nginx can be run on something as simple as a raspberry PI.  I run it on a network file
server attached to my desktop via copper ethernet.  By caching the footprint libraries
from github, I can load all github libraries in about 2.5 seconds, making github the
fastest plugin of any kind.