Make pagelayout_default.kicad_wks match defaultPageLayout
The pagelayout_default.kicad_wks template differs slightly from
common/page_layout/page_layout_default_description.cpp:defaultPageLayout

Not sure if this is intentional or a bug. In any case, here is a
patch that changes pagelayout_default.kicad_wks to match
defaultPageLayout.

This patch is based on the previous one that added missing
parentheses.