wxWidgets OSX Sierra 10.12 SDK patch.
OSX Sierra has removed QuickTime entirely (because Apple needs to break
at least one thing every major OS revision).  This prevents wxWidgets
from compiling at all, and as of right now, there is only a workaround,
but a true upstream fix will require rewriting portions of wx to use
AVKit, which is more or less the replacement for QuickTime, instead.

KiCad doesn't use any parts of wx involved, so the work around seems to
be fine for us.  I am using the latest commit of KiCad on 10.12 as I write
this, with this patch applied, and I'm not having any problems (at least
new ones anyway).

The workaround is this patch (which is not a direct copy of the upstream
fix, which simply removes the include entirely.  I made it where if this
patch gets applied, it won't effect anything except when being built
against the 10.12 SDK).  Additionally, wxWidgets configure must be given
the --disable-mediactrl flag.  This is a temporary workaround, but as far
as I know, KiCad doesn't use that control?