Add wxWidgets bug fix patch.
Calling external applications with wxWidgets failed when the file name
contained non-ASCII characters.

Please note that this patch does not require wxWidgets to be recompiled.
It only changes a header file which means the patch can be applied to
an existing wxWidgets install.  However, KiCad does need to be rebuilt
after this patch is applied.  This patch is not required if you do not
need non-ASCII character file name support.